climate-impact-forecasting/climate-impact-forecasting/requirements.txt

tensorflow
pandas
geopandas
scikit-learn
flask
paho-mqtt
plotly
streamlit
joblib
